moved more things to widgets.py
