+++ kmpc/resources/interface.kv
@@ -5,11 +5,11 @@
 #:import PlaylistRow kmpc.playlistpanel.PlaylistRow
 #:import PlaylistRecycleBoxLayout kmpc.playlistpanel.PlaylistRecycleBoxLayout
 #:import LibraryTabbedPanelItem kmpc.librarypanel.LibraryTabbedPanelItem
-#:import ExtraSlider kmpc.extra.ExtraSlider
-#:import ClearButton kmpc.extra.ClearButton
-#:import OutlineLabel kmpc.extra.OutlineLabel
-#:import OutlineButton kmpc.extra.OutlineButton
-#:import OutlineTabbedPanelItem kmpc.extra.OutlineTabbedPanelItem
+#:import ExtraSlider kmpc.widgets.ExtraSlider
+#:import ClearButton kmpc.widgets.ClearButton
+#:import OutlineLabel kmpc.widgets.OutlineLabel
+#:import OutlineButton kmpc.widgets.OutlineButton
+#:import OutlineTabbedPanelItem kmpc.widgets.OutlineTabbedPanelItem
 #:import InfoLargeLabel kmpc.widgets.InfoLargeLabel
 
 <OutlineLabel>:
